% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRegularizedModel.R
\name{computeOptimizationStatistics}
\alias{computeOptimizationStatistics}
\title{computeOptimizationStatistics}
\usage{
computeOptimizationStatistics(model, numStats, useSubmodels, saturatedDoF,
  independenceDoF, retval)
}
\description{
This is overridden from OpenMx.  This function is exported for people
   who already know what they are doing.
}
