% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeMxModel.R
\name{getParamsInMatrix}
\alias{getParamsInMatrix}
\title{getParamsInMatrix}
\usage{
getParamsInMatrix(model, matrix)
}
\arguments{
\item{model}{the MxModel}

\item{matrix}{The matrix from which to draw all associated}
}
\value{
a vector containing the names of parameters in a given matrix in the model
}
\description{
Get all free parameters from a given matrix
}
