% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRegularizedModel.R
\name{mxModel}
\alias{mxModel}
\title{Create a (regularizable) mxModel}
\usage{
mxModel(
  model = NA,
  ...,
  manifestVars = NA,
  latentVars = NA,
  remove = FALSE,
  independent = NA,
  type = NA,
  name = NA
)
}
\description{
Identical to OpenMx::mxModel, except the result is an MxRegularizedModel.
}
