% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPenaltySearch.R
\name{mxPenaltySearch}
\alias{mxPenaltySearch}
\title{Search for an optimal penalty value  **Experimental**}
\usage{
mxPenaltySearch(model, search_space = NULL, epsilon = 1e-06, ...)
}
\arguments{
\item{model}{an MxRegularizedModel object to regularize}

\item{...}{the search space for this sequence in format *param=c(...)* where c(...) is the list of possible values for parameter param.}
}
\value{
A regularized MxModel; at this point an MxRegularizedModel
}
\description{
Perform a grid search over a space of regularization penalties.
Currently only guaranteed to work for single-parameter models.
Expect many changes in the interface while this is developed further.
}
