% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPenaltySearch.R
\name{mxPenaltySearchExternal}
\alias{mxPenaltySearchExternal}
\title{External search for an optimal penalty value}
\usage{
mxPenaltySearchExternal(model, search_space = NULL, epsilon = 1e-06,
  ..., approach = "EBIC", ebicGamma = 0.5, returnConstrained = FALSE,
  verbose = interactive())
}
\arguments{
\item{model}{an MxRegularizedModel object to regularize}

\item{epsilon}{how close to zero is zero?}

\item{...}{the search space for this sequence in format *param=c(...)* where c(...) is the list of possible values for parameter param.}

\item{approach}{what fit function to use?  Currently only EBIC is available}

\item{ebicGamma}{what Gamma value to use for EBIC?  Must be between 0 and 1}

\item{returnConstrained}{fix all regularized-to-zero values to zero, and rerun fit without regularization}
}
\value{
A regularized MxModel; at this point an MxRegularizedModel
}
\description{
Perform a grid search over a space of regularization penalties.
This happens in the front-end, returning to R at each cycle.
Currently only guaranteed to work for single-parameter models.
Expect many changes in the interface while this is developed further.
**WARNING: Dangerously assumes that fitfunction is the -2 log likelihood**
}
\details{
The EBIC computation uses the computation from the previous regsem.
It is not clear why this is the case.

TODO: Figure out how to return all the individual runs.
}
