% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPenalty.R
\name{mxRegularize}
\alias{mxRegularize}
\title{This function creates a regularization penalty object}
\usage{
mxRegularize(what, how = names(imxRegularizationTypes),
  hyperparams = list(), hpranges = list(), name = NULL,
  compute = mxComputeDefault(), otherArgs = list())
}
\arguments{
\item{what}{which parameters or model elements to regularize}

\item{how}{what kind of regularization function to use.  Currently supported: "lasso", "ridge"}

\item{hyperparams}{a named list of the hyperparameter starting values. Used for initial setup.}

\item{hpranges}{a named list of hyperparameter ranges. Used in search if no ranges are specified.}

\item{name}{the name of the regularization object to be created}

\item{compute}{a compute object.  If it contains at least one mxComputeDefault, any compute plan supplied by the user will be substituted for that object each time it occurs in the plan.}

\item{other}{Other arguments to the associated regularization functions (see mxRegularizeLASSO and mxRegularizeRidge)}
}
\description{
This function creates a regularization penalty object
}
