% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPenaltyFunctions.R
\name{mxRegularizeConstant}
\alias{mxRegularizeConstant}
\title{MxRegularizeConstant}
\usage{
mxRegularizeConstant(name, penalty)
}
\arguments{
\item{name}{Name of the regularizer object}

\item{penalty}{How much to penalize}
}
\description{
Constant regularization penalty for no good reason
}
