% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPenaltyFunctions.R
\name{mxRegularizeElasticNet}
\alias{mxRegularizeElasticNet}
\title{MxRegularizeElasticNet}
\usage{
mxRegularizeElasticNet(what, name, alpha = 0, alpha.step = 0.1,
  alpha.max = 1, lambda = 0, lambda.step = 0.1, lambda.max = 0.4,
  alpha.min = NA, lambda.min = NA)
}
\arguments{
\item{what}{A list of parameters to regularize}

\item{name}{Name of the regularizer object}

\item{alpha}{strength of the mixing parameter to be applied at start (default 0.5).  Note that 0 indicates a ridge regression with penalty \deqn{\frac{lambda}{2}}{lambda / 2}, and 1 indicates a LASSO regression with penalty lambda.}

\item{alpha.step}{alpha step during penalty search (default 0.1)}

\item{alpha.max}{when to end the alpha search (default 1)}

\item{lambda}{strength of the penalty to be applied at starting values (default 0)}

\item{lambda.step}{step function for lambda step (default .01)}

\item{lambda.max}{end of lambda range (default .4)}

\item{alpha.min}{beginning of the alpha range (default 0)}

\item{lambda.min}{beginning of the lambda range (default lambda)}
}
\description{
Elastic net regularization
}
\details{
Applies elastic net regularization.  Elastic net is a weighted combination of ridge and LASSO penalties.
}
