% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxPenaltyFunctions.R
\name{mxRegularizeRidge}
\alias{mxRegularizeRidge}
\title{MxRegularizeRidge}
\usage{
mxRegularizeRidge(
  what,
  name,
  lambda = 0,
  lambda.step = 0.01,
  lambda.max = 0.4,
  lambda.min = NA
)
}
\arguments{
\item{what}{A list of parameters to regularize}

\item{name}{Name of the regularizer object}

\item{lambda}{strength of the penalty to be applied at start (default 0)}

\item{lambda.step}{lambda step during penalty search (default 0.01)}

\item{lambda.max}{when to end the lambda search (default 0.4)}

\item{lambda.min}{minimum lambda value (default lambda)}
}
\description{
Ridge regression regularization
}
