% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReg.R
\name{plotReg}
\alias{plotReg}
\title{Plotting function to see some of the regularization outputs cleanly}
\usage{
plotReg(model)
}
\arguments{
\item{model}{an MxRegularizedModel object to regularize}
}
\value{
None
}
\description{
Plotting function to see some of the regularization outputs cleanly
}
\details{
Plots of the regularization tradeoffs
}
