% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeMxModel.R
\name{regularizeMxModel}
\alias{regularizeMxModel}
\title{Regularize an MxModel}
\usage{
regularizeMxModel(
  model,
  reg_params = NULL,
  penalty_value = 50,
  penalty_function = "lasso"
)
}
\arguments{
\item{model}{the original mxmodel to regularize}

\item{reg_params}{the list of parameters or algebra cells regularized by this object.  The default (NULL) regularizes on all free parameters in the model.}

\item{penalty_function}{the type of function to eb used for regularization.  Currently supports "lasso" or "ridge"}

\item{...}{arguments to the penalty function, usually `lambda=` (for lasso or ridge) and `alpha=` (for elastic net)}
}
\value{
A regularized MxModel; at this point a wrapper model with regularization enclosed
}
\description{
Create a regularized container object that encloses an unregularized mxModel.
This was the first trial; you should now just create MxModels using MxModel.
}
