% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeMxModel.R
\name{summarizeRegularized}
\alias{summarizeRegularized}
\title{Regularize an MxModel}
\usage{
summarizeRegularized(regFit, ..., verbose = FALSE,
  penalty_function = "guess", epsilon = 1e-06)
}
\arguments{
\item{...}{values passed to mxSummarize}

\item{verbose}{gives more summary feedback}

\item{penalty_function}{the type of function used for regularization. Default guesses from regularization algebra name}

\item{epsilon}{how close to zero is close enough? Regularization limits for DF correction}

\item{model}{the regularized model to summarize}
}
\value{
Summary object summarizing the mxModel
}
\description{
Create a regularized container object that encloses an unregularized mxModel.
}
