% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRegularizedModel.R
\name{summary,MxRegularizedModel-method}
\alias{summary,MxRegularizedModel-method}
\title{Summarize an MxRegularizedModel}
\usage{
\S4method{summary}{MxRegularizedModel}(object, ..., verbose = FALSE, computeRefs = FALSE, epsilon = 1e-06)
}
\arguments{
\item{...}{values passed to mxSummarize}

\item{verbose}{gives more summary feedback}

\item{epsilon}{how close to zero is close enough? Regularization limits for DF correction}

\item{regFit}{the regularized model to summarize}
}
\value{
Summary object summarizing the MxRegularizedModel
}
\description{
Create a summary of an MxRegularizedModel.
}
